Wrote additional javadoc in api.API
Added the api.options package
Added and wrote the api.options.AbstractOption class, a superclass for
options (later on --verbose, --id <id> and many more).

Added and wrote the api.options.AbstractParameteredOption class, a generic
superclass for options with a parameter (later on --id <id> and many
more).

The Abstract*Option classes are written to pass a set of options (later on
conditions) for the delete, export and list commands and will be very
useful in the future.

STILL MISSING javadoc for delete, export, list and info methods in api.API